Version 0.1.0
- Now requires Laser Utils 5.3.1
  - HUD Overlays are now registered using LUOverlayManager

- All player data is now managed under one capability

- Removed ManageHuntersCommand
- Added SetPlayerRoleCommands to properly manage the roles of players for a Manhunt game

- Added title and subtitles for when the Manhunt game starts, pauses, and resumes

- Users can now save Manhunt settings to a json file and start a Manhunt game using the settings from a json file

- Added SpeedRunnerToSpectatorEvent
  - Called when a player becomes a spectator

- Fixed Tracker tracking players that switched to spectator
