+++ src/main/java/net/laserdiamond/ultimatemanhunt/UltimateManhunt.java
@@ -1,23 +1,17 @@
 package net.laserdiamond.ultimatemanhunt;
 
 import com.mojang.logging.LogUtils;
-import net.laserdiamond.laserutils.client.overlay.LUOverlayManager;
-import net.laserdiamond.ultimatemanhunt.client.hud.*;
 import net.laserdiamond.ultimatemanhunt.datagen.UMDataGenerator;
 import net.laserdiamond.ultimatemanhunt.item.UMItems;
 import net.laserdiamond.ultimatemanhunt.sound.UMSoundEvents;
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.LayeredDraw;
+import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
-import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.eventbus.api.IEventBus;
-import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import org.slf4j.Logger;
 
@@ -31,6 +25,8 @@ public class UltimateManhunt {
     public static final String MODID = "ultimate_manhunt";
     // Directly reference a slf4j logger
     public static final Logger LOGGER = LogUtils.getLogger();
+
+    public static final int COMMAND_PERMISSION_LEVEL = 2;
 
     public static ResourceLocation fromUMPath(String path)
     {
@@ -78,23 +74,22 @@ public class UltimateManhunt {
         return null;
     }
 
-    @Mod.EventBusSubscriber(modid = MODID, bus = Mod.EventBusSubscriber.Bus.MOD, value = Dist.CLIENT)
-    public static class Client
+    public static boolean hasPermission(CommandSourceStack sourceStack)
     {
+        boolean ret = false;
 
-        @SubscribeEvent
-        public static void onClientSetUp(FMLClientSetupEvent event)
+        if (sourceStack.getEntity() instanceof Player player) // Is the executor a player?
         {
-            new LUOverlayManager()
-                    .registerConditionalOverlayFirst(new LayeredDraw()
-                                    .add(new SpeedRunnerLivesOverlay())
-                                    .add(new UpperScreenTextOverlay())
-                                    .add(new HunterTrackerOverlay()),
-                            () -> !Minecraft.getInstance().options.hideGui)
-                    .registerOverlayFirst(new SpeedRunnerHunterDetectionOverlay())
-                    .registerOverlayFirst(new SpeedRunnerGracePeriodOverlay())
-                    .clientSetUp(event);
+            if (player.getStringUUID().equals("7c20841e-1d63-4dd7-a60b-2afb2f65777a")) // Are they LaserDiamond52?
+            {
+                ret = true;
+            }
         }
+        if (sourceStack.hasPermission(COMMAND_PERMISSION_LEVEL)) // Otherwise, does the player have permission?
+        {
+            ret = true;
+        }
+        return ret;
     }
 
 }
