+++ src/main/java/net/laserdiamond/ultimatemanhunt/commands/gamerule/SetWindTorchesCommand.java
@@ -4,7 +4,7 @@ import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.arguments.BoolArgumentType;
 import com.mojang.brigadier.context.CommandContext;
 import net.laserdiamond.ultimatemanhunt.UMGame;
-import net.laserdiamond.ultimatemanhunt.event.ForgeServerEvents;
+import net.laserdiamond.ultimatemanhunt.UltimateManhunt;
 import net.minecraft.ChatFormatting;
 import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.commands.Commands;
@@ -12,13 +12,11 @@ import net.minecraft.network.chat.Component;
 
 public class SetWindTorchesCommand {
 
-    private static final int PERMISSION_LEVEL = 2;
-
     public static void register(CommandDispatcher<CommandSourceStack> dispatcher)
     {
         dispatcher.register(
                 Commands.literal("um_wind_torches")
-                        .requires(sourceStack -> ForgeServerEvents.permission(sourceStack, PERMISSION_LEVEL))
+                        .requires(UltimateManhunt::hasPermission)
                         .then(
                                 Commands.argument("enableWindTorches", BoolArgumentType.bool())
                                         .executes(commandContext -> setWindTorches(commandContext, BoolArgumentType.getBool(commandContext, "enableWindTorches")))
